Replace use of deprecated numpy aliases for python int, float, bool.

+++ astropy/coordinates/tests/test_matching.py
@@ -194,12 +194,12 @@ def test_search_around():
     coo1 = ICRS([4.1, 2.1]*u.degree, [0, 0]*u.degree, distance=[1, 5] * u.kpc)
     idx1, idx2, d2d, d3d = search_around_sky(coo1, coo2, 1*u.arcsec)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
     idx1, idx2, d2d, d3d = search_around_3d(coo1, coo2, 1*u.m)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
 
@@ -207,33 +207,33 @@ def test_search_around():
     empty = ICRS(ra=[] * u.degree, dec=[] * u.degree, distance=[] * u.kpc)
     idx1, idx2, d2d, d3d = search_around_sky(empty, coo2, 1*u.arcsec)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
     idx1, idx2, d2d, d3d = search_around_sky(coo1, empty, 1*u.arcsec)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
     empty = ICRS(ra=[] * u.degree, dec=[] * u.degree, distance=[] * u.kpc)
     idx1, idx2, d2d, d3d = search_around_sky(empty, empty[:], 1*u.arcsec)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
     idx1, idx2, d2d, d3d = search_around_3d(empty, coo2, 1*u.m)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
     idx1, idx2, d2d, d3d = search_around_3d(coo1, empty, 1*u.m)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
     idx1, idx2, d2d, d3d = search_around_3d(empty, empty[:], 1*u.m)
     assert idx1.size == idx2.size == d2d.size == d3d.size == 0
-    assert idx1.dtype == idx2.dtype == np.int
+    assert idx1.dtype == idx2.dtype == int
     assert d2d.unit == u.deg
     assert d3d.unit == u.kpc
